Add files via upload
It would seem I was trying to delete objects from within nested functions.  I made a boolean active property for balls, and it is all good.  active flips false when the ball hits the bottom line (currently at height 270) and then I use that property to delete the object with splice.  When all is said and done, balls[] is empty, which is what I wanted.  Next up is the lead property to minimize collision calculations.