Add files via upload
which is faster?  using a quadruple or statement to decide individually check and act on each if statement, or just doing the if statements stand alone.  I time trialed it, don't need the gate keeper quadruple or statement.  So I worked out some stuff and now Im into the kinks of flipping the exposed boolean for the blocks.  more fun times ahead, but I'm too tired now to be productive.