Add files via upload
Yes this is a new document name.  I'll sort all that out when I have time.  So, I thought about some things during idle time and I figured I had the collision detection and bouncing all wrong.  So in a few hours of chaotic typing and debugging, I got it sorted.  all of it.  turns out the intercepts funciton I copied from StackOverflow was slightly flawed for my uses.  It wouldn't detect if the end point of one line was on another line.  TL;DR Fixed that, restructured the collision and everything works like a charm.  Now i just need to start on stage construciton.